improved it a little  more
