+++ cmd/kusk/cmd/manifests/dashboard.yaml
@@ -1,0 +1,92 @@
+apiVersion: v1
+kind: ServiceAccount
+metadata:
+  name: kusk-gateway-dashboard
+  namespace: kusk-system
+---
+apiVersion: v1
+kind: Service
+metadata:
+  name: kusk-gateway-dashboard
+  namespace: kusk-system
+spec:
+  ports:
+    - name: http
+      port: 80
+      protocol: TCP
+      targetPort: http
+  selector:
+    app.kubernetes.io/instance: kusk-gateway-dashboard
+    app.kubernetes.io/name: kusk-gateway-dashboard
+  type: ClusterIP
+---
+apiVersion: apps/v1
+kind: Deployment
+metadata:
+  name: kusk-gateway-dashboard
+  namespace: kusk-system
+  labels:
+    app.kubernetes.io/instance: kusk-gateway-dashboard
+    app.kubernetes.io/name: kusk-gateway-dashboard
+    app.kubernetes.io/version: v1.1.0-stable
+spec:
+  selector:
+    matchLabels:
+      app.kubernetes.io/instance: kusk-gateway-dashboard
+      app.kubernetes.io/name: kusk-gateway-dashboard
+  strategy:
+    rollingUpdate:
+      maxSurge: 25%
+      maxUnavailable: 25%
+    type: RollingUpdate
+  template:
+    metadata:
+      labels:
+        app.kubernetes.io/instance: kusk-gateway-dashboard
+        app.kubernetes.io/name: kusk-gateway-dashboard
+    spec:
+      containers:
+        - image: kubeshop/kusk-gateway-dashboard:v1.1.0-stable
+          imagePullPolicy: Always
+          livenessProbe:
+            httpGet:
+              path: /
+              port: http
+              scheme: HTTP
+          name: kusk-gateway-dashboard
+          ports:
+            - containerPort: 80
+              name: http
+              protocol: TCP
+          readinessProbe:
+            httpGet:
+              path: /
+              port: http
+              scheme: HTTP
+      serviceAccount: kusk-gateway-dashboard
+      serviceAccountName: kusk-gateway-dashboard
+---
+apiVersion: gateway.kusk.io/v1alpha1
+kind: StaticRoute
+metadata:
+  name: kusk-gateway-dashboard
+  namespace: kusk-system
+spec:
+  fleet:
+    name: kusk-gateway-private-envoy-fleet
+    namespace: kusk-system
+  hosts:
+  - localhost
+  - '*'
+  paths:
+    /:
+      get:
+        route:
+          upstream:
+            rewrite:
+              pattern: ""
+              substitution: ""
+            service:
+              name: kusk-gateway-dashboard
+              namespace: kusk-system
+              port: 80
