+++ cmd/kusk/cmd/manifests/fleets.yaml
@@ -1,0 +1,41 @@
+apiVersion: gateway.kusk.io/v1alpha1
+kind: EnvoyFleet
+metadata:
+  name: kusk-gateway-envoy-fleet
+  namespace: kusk-system
+spec:
+  accesslog:
+    format: text
+  image: envoyproxy/envoy:v1.23.1
+  service:
+    ports:
+      - name: http
+        port: 80
+        protocol: TCP
+        targetPort: http
+      - name: https
+        port: 443
+        protocol: TCP
+        targetPort: http
+    type: LoadBalancer
+---
+apiVersion: gateway.kusk.io/v1alpha1
+kind: EnvoyFleet
+metadata:
+  name: kusk-gateway-private-envoy-fleet
+  namespace: kusk-system
+spec:
+  accesslog:
+    format: text
+  image: envoyproxy/envoy:v1.23.1
+  service:
+    ports:
+      - name: http
+        port: 80
+        protocol: TCP
+        targetPort: http
+      - name: https
+        port: 443
+        protocol: TCP
+        targetPort: http
+    type: ClusterIP
